requirements.txt
importlib
bisect
pickle
zipfile
pandas
numpy
matplotlib
streamlit
streamlit_image_coordinates
scikit-image
opencv-python
pillow